NavBar Refactoring
